+++ packages/bpk-component-fieldset/stories.js
@@ -6,7 +6,7 @@ import BpkInput, { INPUT_TYPES } from 'bpk-component-input';
 
 import BpkFieldset from './index';
 
-class ValidationContainer extends Component {
+class FieldsetContainer extends Component {
   constructor() {
     super();
 
@@ -26,14 +26,7 @@ class ValidationContainer extends Component {
   }
 
   render() {
-    const { ...rest } = this.props;
-
-    delete rest.value;
-    delete rest.valid;
-    delete rest.validValue;
-    delete rest.onChange;
-
-    const isCheckbox = rest.control.name === 'BpkCheckbox';
+    const { validValue, isCheckbox, ...rest } = this.props;
 
     const valueProps = isCheckbox
       ? { checked: this.state.checked }
@@ -41,16 +34,17 @@ class ValidationContainer extends Component {
 
     let isValid;
     if (isCheckbox) {
-      isValid = this.state.checked === this.props.validValue;
+      isValid = this.state.checked === validValue;
     } else {
       isValid = this.state.value === ''
         ? undefined
-        : this.state.value === this.props.validValue;
+        : this.state.value === validValue;
     }
 
     return (
       <BpkFieldset
         valid={isValid}
+        isCheckbox={isCheckbox}
         onChange={this.onChange}
         {...rest}
         {...valueProps}
@@ -59,16 +53,21 @@ class ValidationContainer extends Component {
   }
 }
 
-ValidationContainer.propTypes = {
+FieldsetContainer.propTypes = {
   validValue: PropTypes.oneOfType([
     PropTypes.string,
     PropTypes.bool,
   ]).isRequired,
+  isCheckbox: PropTypes.bool,
+};
+
+FieldsetContainer.defaultProps = {
+  isCheckbox: false,
 };
 
 storiesOf('bpk-component-fieldset', module)
   .add('Input example', () => (
-    <ValidationContainer
+    <FieldsetContainer
       id="name_input"
       name="name"
       label="Name"
@@ -80,7 +79,7 @@ storiesOf('bpk-component-fieldset', module)
     />
   ))
   .add('Select example', () => (
-    <ValidationContainer
+    <FieldsetContainer
       id="fruits_select"
       name="fruits"
       label="Fruits"
@@ -93,15 +92,16 @@ storiesOf('bpk-component-fieldset', module)
       <option value="oranges">Oranges</option>
       <option value="pears">Pears</option>
       <option value="tomato" disabled>Tomato</option>
-    </ValidationContainer>
+    </FieldsetContainer>
   ))
   .add('Checkbox example', () => (
-    <ValidationContainer
+    <FieldsetContainer
       id="prefer_directs_checkbox"
       name="prefer_directs"
       label="Prefer directs"
       control={BpkCheckbox}
       validationMessage="Required"
       validValue
+      isCheckbox
     />
   ));