+++ packages/bpk-component-form-validation/src/BpkFormValidation-test.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import renderer from 'react-test-renderer';
+
+import BpkFormValidation from './BpkFormValidation';
+
+describe('BpkFieldset', () => {
+  it('should render correctly', () => {
+    const tree = renderer.create(
+      <BpkFormValidation />,
+    ).toJSON();
+    expect(tree).toMatchSnapshot();
+  });
+});