+++ packages/bpk-component-form-validation/stories.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import { storiesOf } from '@kadira/storybook';
+
+import BpkFormValidation from './index';
+
+storiesOf('bpk-component-form-validation', module)
+  .add('Default', () => (
+    <BpkFormValidation>
+      TODO
+    </BpkFormValidation>
+  ));